+++ client/highlevel-rest/licenses/commons-codec-1.10.jar.sha1
