+++ client/highlevel-rest/licenses/jackson-core-2.8.1.jar.sha1
