+++ settings.gradle
@@ -6,7 +6,6 @@ List projects = [
   'core',
   'docs',
   'client:rest',
-  'client:advanced-rest',
   'client:sniffer',
   'client:transport',
   'client:highlevel-rest',
